+++ packages/react-eyecandy/tsup.config.ts
@@ -2,9 +2,8 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig({
   entry: ['src'],
-  format: ['cjs'],
-  splitting: false,
   sourcemap: true,
   clean: true,
   dts: true,
+  outDir: 'lib',
 });