+++ packages/react-eyecandy/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @nomada-sh/react-eyecandy
 
+## 0.0.3
+
+### Patch Changes
+
+- Using `react-css-theme-switcher` to change css files at runtime.
+- Changed outDir to `lib`.
+
 ## 0.0.2
 
 ### Patch Changes
 
-- Added lib to files in package.json
-- Added antd as a peerDependency
+- Added `lib` to files in package.json
+- Added `antd` as a peerDependency
 
 ## 0.0.1